packages/py-try/try-django/pyproject.toml
[tool.poetry]
name = "try-django"
version = "0.1.0"
description = ""
authors = ["Henry.Huang <hhglory@outlook.com>"]
readme = "README.md"
packages = [{include = ".", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.0"
sentry-sdk = "^1.39.1"
django-debug-toolbar = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
